Cleanup pass: clippy hygiene, doc trims, dead-consumer refs
Clippy:
- Prune 17 unused imports left over from the const-num-traits migration.
- Two `assign_op_pattern` fixes (`|=`) inside c0nst impls.
- `#![allow(clippy::let_unit_value)]` on `byte_conversion_panic_free`
  with a one-line rationale — the bare-statement suggestion breaks
  const-eval of the `AssertBufferFits<M>::CHECK` assertion.
- `#[allow(clippy::op_ref)]` on 3 test modules that deliberately cover
  every ref/value combination of the operator impls.
- Cfg-gate `hex_fmt`, `make_overflow_err`, `make_empty_error`,
  `use core::fmt::Write` behind `feature = "num-traits"` so
  `--no-default-features` builds without unused-code warnings.
- Various small test-side fixes (`u32::max_value()` -> `u32::MAX`,
  `vec!` -> `[..]`, unused type parameters flagged, etc.).

Result: clippy clean under all feature combos across `--lib --tests`.

Docs:
- Drop specific downstream consumer name-drops (`modmath`,
  `ed25519_heapless`, etc.) from tracked prose. Actual code references
  to `modmath-cios` (the trait crate) stay; only the "consumer X
  will use this" narrative goes.
- Drop migration-era `Phase 3c/d/e/...` markers from ct-fixtures
  section headers and `Phase 1/2 row` labels on CIOS methods.
- Trim four novella-length module doc blocks
  (byte_conversion_panic_free, cios_row_ops_impl, has_nonzero_impl,
  power_of_two_ops_impl) down to the essentials. Net 224 lines of
  prose removed.
- Drop the stale `const_trait_impl` TODO comment in machineword.rs;
  the `c0nst!` macro already renders that trait as const on nightly.

+++ tests/bit_shift.rs
@@ -265,7 +265,7 @@ fn test_rotate() {
         assert_eq!(b.rotate_right(full_shift), b);
 
         // Rotations larger than the bit width should wrap around
-        let bit_width = INT::BIT_SIZE as u32;
+        let bit_width = INT::BIT_SIZE;
         let overflow_shift = full_shift + 5;
         let expected_left = a.rotate_left(overflow_shift % bit_width);
         assert_eq!(a.rotate_left(overflow_shift), expected_left);
